extract <entries> component
